requirements_prod.txt
asgiref==3.5.2
Django==4.1
django-countries==7.3.2
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
pydantic==1.9.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
gunicorn==20.1.0
